+++ extensions/api/data-management/api-configuration/src/main/java/org/eclipse/dataspaceconnector/api/datamanagement/configuration/DataManagementApiConfigurationExtension.java
@@ -24,6 +24,7 @@ import org.eclipse.dataspaceconnector.spi.system.ServiceExtension;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
 
 import static java.lang.String.format;
+import static java.util.Optional.ofNullable;
 
 @Provides(DataManagementApiConfiguration.class)
 public class DataManagementApiConfigurationExtension implements ServiceExtension {
@@ -33,7 +34,8 @@ public class DataManagementApiConfigurationExtension implements ServiceExtension
 
     @Inject
     private WebService webService;
-    @Inject
+
+    @Inject(required = false)
     private AuthenticationService service;
 
     @Override
@@ -61,7 +63,8 @@ public class DataManagementApiConfigurationExtension implements ServiceExtension
         // the DataManagementApiConfiguration tells all DataManagementApi controllers under which context alias
         // they need to register their resources: either `default` or `data`
         context.registerService(DataManagementApiConfiguration.class, new DataManagementApiConfiguration(contextAlias));
-        webService.registerResource(contextAlias, new AuthenticationRequestFilter(service));
+        var srv = ofNullable(service).orElse(headers -> true);
+        webService.registerResource(contextAlias, new AuthenticationRequestFilter(srv));
         webService.registerResource(contextAlias, new EdcApiExceptionMapper());
     }
 }