+++ extensions/api/data-management/policydefinition/src/main/java/org/eclipse/dataspaceconnector/api/datamanagement/policy/PolicyApiController.java
@@ -24,13 +24,19 @@ import jakarta.ws.rs.PathParam;
 import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.QueryParam;
 import jakarta.ws.rs.core.MediaType;
-import org.eclipse.dataspaceconnector.api.datamanagement.policy.model.PolicyDefinitionDto;
+import org.eclipse.dataspaceconnector.api.datamanagement.policy.service.PolicyService;
+import org.eclipse.dataspaceconnector.api.exception.ObjectExistsException;
+import org.eclipse.dataspaceconnector.api.exception.ObjectNotFoundException;
+import org.eclipse.dataspaceconnector.api.result.ServiceResult;
+import org.eclipse.dataspaceconnector.policy.model.Policy;
+import org.eclipse.dataspaceconnector.spi.EdcException;
 import org.eclipse.dataspaceconnector.spi.monitor.Monitor;
 import org.eclipse.dataspaceconnector.spi.query.QuerySpec;
 import org.eclipse.dataspaceconnector.spi.query.SortOrder;
 
-import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 import static java.lang.String.format;
 
@@ -38,19 +44,22 @@ import static java.lang.String.format;
 @Produces({ MediaType.APPLICATION_JSON })
 @Path("/policies")
 public class PolicyApiController implements PolicyApi {
-    private final Monitor monitor;
 
-    public PolicyApiController(Monitor monitor) {
+    private final Monitor monitor;
+    private final PolicyService policyService;
+
+    public PolicyApiController(Monitor monitor, PolicyService policyService) {
         this.monitor = monitor;
+        this.policyService = policyService;
     }
 
     @GET
     @Override
-    public List<PolicyDefinitionDto> getAllPolicies(@QueryParam("offset") Integer offset,
-                                                    @QueryParam("limit") Integer limit,
-                                                    @QueryParam("filter") String filterExpression,
-                                                    @QueryParam("sort") SortOrder sortOrder,
-                                                    @QueryParam("sortField") String sortField) {
+    public List<Policy> getAllPolicies(@QueryParam("offset") Integer offset,
+                                       @QueryParam("limit") Integer limit,
+                                       @QueryParam("filter") String filterExpression,
+                                       @QueryParam("sort") SortOrder sortOrder,
+                                       @QueryParam("sortField") String sortField) {
         var spec = QuerySpec.Builder.newInstance()
                 .offset(offset)
                 .limit(limit)
@@ -59,23 +68,32 @@ public class PolicyApiController implements PolicyApi {
                 .sortOrder(sortOrder).build();
         monitor.debug(format("get all policys %s", spec));
 
-        return Collections.emptyList();
+        return policyService.query(spec).stream()
+                .collect(Collectors.toList());
 
     }
 
     @GET
     @Path("{id}")
     @Override
-    public PolicyDefinitionDto getPolicy(@PathParam("id") String id) {
-        monitor.debug(format("get policy with ID %s", id));
-
-        return null;
+    public Policy getPolicy(@PathParam("id") String id) {
+        monitor.debug(format("Attempting to return policy with ID %s", id));
+        return Optional.of(id)
+                .map(it -> policyService.findById(id))
+                .orElseThrow(() -> new ObjectNotFoundException(Policy.class, id));
     }
 
     @POST
     @Override
-    public void createPolicy(PolicyDefinitionDto dto) {
-        monitor.debug("create new policy");
+    public void createPolicy(Policy policy) {
+
+        var result = policyService.create(policy);
+
+        if (result.succeeded()) {
+            monitor.debug(format("Policy created %s", policy.getUid()));
+        } else {
+            handleFailedResult(result, policy.getUid());
+        }
     }
 
     @DELETE
@@ -83,6 +101,23 @@ public class PolicyApiController implements PolicyApi {
     @Override
     public void deletePolicy(@PathParam("id") String id) {
         monitor.debug(format("Attempting to delete policy with id %s", id));
+        var result = policyService.deleteById(id);
+        if (result.succeeded()) {
+            monitor.debug(format("Policy deleted %s", id));
+        } else {
+            handleFailedResult(result, id);
+        }
+    }
+
+    private void handleFailedResult(ServiceResult<Policy> result, String id) {
+        switch (result.reason()) {
+            case NOT_FOUND:
+                throw new ObjectNotFoundException(Policy.class, id);
+            case CONFLICT:
+                throw new ObjectExistsException(Policy.class, id);
+            default:
+                throw new EdcException("unexpected error");
+        }
     }
 
 }