+++ extensions/api/data-management/transferprocess/src/main/java/org/eclipse/dataspaceconnector/api/datamanagement/transferprocess/service/TransferProcessService.java
@@ -16,7 +16,7 @@ package org.eclipse.dataspaceconnector.api.datamanagement.transferprocess.servic
 
 import org.eclipse.dataspaceconnector.api.result.ServiceResult;
 import org.eclipse.dataspaceconnector.spi.query.QuerySpec;
-import org.eclipse.dataspaceconnector.spi.result.Result;
+import org.eclipse.dataspaceconnector.spi.types.domain.transfer.DataRequest;
 import org.eclipse.dataspaceconnector.spi.types.domain.transfer.TransferProcess;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -64,7 +64,7 @@ public interface TransferProcessService {
      * @return a result that is successful if the transfer process was found and is in a state that can be canceled
      */
     @NotNull
-    ServiceResult<?> cancel(String transferProcessId);
+    ServiceResult<TransferProcess> cancel(String transferProcessId);
 
     /**
      * Asynchronously requests deprovisioning of the transfer process.
@@ -75,5 +75,14 @@ public interface TransferProcessService {
      * @return a result that is successful if the transfer process was found and is in a state that can be deprovisioned
      */
     @NotNull
-    ServiceResult<?> deprovision(String transferProcessId);
+    ServiceResult<TransferProcess> deprovision(String transferProcessId);
+
+    /**
+     * Initiate transfer request.
+     *
+     * @param request for the transfer.
+     * @return a result that is successful if the transfer process was initiated with id of created transferProcess.
+     */
+    @NotNull
+    ServiceResult<String> initiateTransfer(DataRequest request);
 }