+++ system-tests/tests/src/test/java/org/eclipse/dataspaceconnector/system/tests/local/FileTransferIntegrationTest.java
@@ -17,6 +17,7 @@
 package org.eclipse.dataspaceconnector.system.tests.local;
 
 import org.eclipse.dataspaceconnector.common.annotations.EndToEndTest;
+import org.eclipse.dataspaceconnector.common.annotations.PerformanceTest;
 import org.eclipse.dataspaceconnector.junit.launcher.EdcRuntimeExtension;
 import org.eclipse.dataspaceconnector.system.tests.utils.FileTransferSimulationUtils;
 import org.junit.jupiter.api.Test;
@@ -37,6 +38,7 @@ import static org.eclipse.dataspaceconnector.system.tests.utils.FileTransferSimu
 import static org.eclipse.dataspaceconnector.system.tests.utils.GatlingUtils.runGatling;
 
 @EndToEndTest
+@PerformanceTest
 public class FileTransferIntegrationTest {
 
     public static final String CONSUMER_ASSET_PATH = new File(tempDirectory(), "output.txt").getAbsolutePath();